requirements.txt
Flask==2.3.2
boto3==1.28.78
Pillow==10.0.1